adopt-a-pal/api/requirements.txt
Flask==2.1.0
google-cloud-datastore
gunicorn==20.1.0
Waitress
python-dotenv